requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.0.7
Flask-Limiter==3.5.1
Flask-Talisman==1.1.0
psycopg2-binary==2.9.11
gunicorn==23.0.0
waitress==3.0.2
python-dotenv==1.2.1
sqlalchemy==2.0.45
werkzeug==3.1.5